backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
pydantic==2.5.3
python-dotenv==1.0.0
transformers==4.37.0
torch==2.1.2
web3==6.14.0
httpx==0.26.0
